+++ tests/src/index.ts
@@ -11,6 +11,7 @@ import * as setMinSnapshotWindow from "./endpoints/set-min-snapshot-window";
 import * as setTokensPerSlot from "./endpoints/set-tokens-per-slot";
 import * as setFarmOwner from "./endpoints/set-farm-owner";
 import * as createFarmer from "./endpoints/create-farmer";
+import * as startFarming from "./endpoints/start-farming";
 
 import { airdrop, provider } from "./helpers";
 
@@ -23,6 +24,7 @@ describe("farming", () => {
   setFarmOwner.test();
   setTokensPerSlot.test();
   createFarmer.test();
+  startFarming.test();
 
   before("airdrop SOL to provider wallet", async () => {
     await airdrop(provider.wallet.publicKey);